+++ lib/slop/option.rb
@@ -159,7 +159,7 @@ class Slop
     # @return [String]
     def inspect
       "#<Slop::Option short_flag=#{@short_flag.inspect} " +
-      "long_flag=#{@long_flag.inspect} " +
+      "long_flag=#{@long_flag.inspect} argument=#{@argument.inspect} " +
       "description=#{@description.inspect}>"
     end